feat(settings): group queue priority boosts, offline queue, and rate-limit hold into Queue Control
Queue Control lands in AI & Processing between Global Defaults and
Segment actions. It collects the process-new-episodes-first toggle and
the three queue priority boosts (moved out of Global Defaults), the
offline queue block (moved out of Data & Security; the section file is
folded in), and a new rate-limit hold block for the 429 pause feature,
including the active-hold banner.
