+++ frontend/src/pages/settings/GlobalDefaultsSection.test.tsx
@@ -7,8 +7,8 @@
  * their own card, SegmentActionsSection; see SegmentActionsSection.test.tsx.
  */
 import { useState } from 'react';
-import { describe, it, expect, vi } from 'vitest';
-import { render, screen, fireEvent } from '@testing-library/react';
+import { describe, it, expect } from 'vitest';
+import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import GlobalDefaultsSection from './GlobalDefaultsSection';
 import type { EpisodeLogLevel, LowAdYieldAction } from '../../api/types';
@@ -28,14 +28,6 @@ function Harness({ onCommit }: { onCommit: (minutes: number) => void }) {
         onMaxFeedEpisodesChange={() => {}}
         onlyExposeProcessedDefault={false}
         onOnlyExposeProcessedDefaultChange={() => {}}
-        processNewEpisodesFirst
-        onProcessNewEpisodesFirstChange={() => {}}
-        queueManualBoost={20}
-        onQueueManualBoostChange={() => {}}
-        queueFreshBoost={5}
-        onQueueFreshBoostChange={() => {}}
-        queueBulkBoost={0}
-        onQueueBulkBoostChange={() => {}}
         lowAdYieldAction="nothing"
         onLowAdYieldActionChange={() => {}}
         episodeLogRetentionDays={30}
@@ -69,14 +61,6 @@ function PodpingHarness({ onCommit }: { onCommit: (payload: PodpingState) => voi
         onMaxFeedEpisodesChange={() => {}}
         onlyExposeProcessedDefault={false}
         onOnlyExposeProcessedDefaultChange={() => {}}
-        processNewEpisodesFirst
-        onProcessNewEpisodesFirstChange={() => {}}
-        queueManualBoost={20}
-        onQueueManualBoostChange={() => {}}
-        queueFreshBoost={5}
-        onQueueFreshBoostChange={() => {}}
-        queueBulkBoost={0}
-        onQueueBulkBoostChange={() => {}}
         lowAdYieldAction="nothing"
         onLowAdYieldActionChange={() => {}}
         episodeLogRetentionDays={30}
@@ -90,103 +74,6 @@ function PodpingHarness({ onCommit }: { onCommit: (payload: PodpingState) => voi
     </>
   );
 }
-
-interface ProcessNewFirstState {
-  processNewEpisodesFirst: boolean;
-}
-
-function ProcessNewFirstHarness({ onCommit }: { onCommit: (payload: ProcessNewFirstState) => void }) {
-  const [processNewEpisodesFirst, setProcessNewEpisodesFirst] = useState(true);
-  return (
-    <>
-      <GlobalDefaultsSection
-        autoProcessEnabled={false}
-        onAutoProcessEnabledChange={() => {}}
-        rssRefreshIntervalMinutes={15}
-        onRssRefreshIntervalMinutesChange={() => {}}
-        podpingEnabled={false}
-        onPodpingEnabledChange={() => {}}
-        maxFeedEpisodes={10}
-        onMaxFeedEpisodesChange={() => {}}
-        onlyExposeProcessedDefault={false}
-        onOnlyExposeProcessedDefaultChange={() => {}}
-        processNewEpisodesFirst={processNewEpisodesFirst}
-        onProcessNewEpisodesFirstChange={setProcessNewEpisodesFirst}
-        queueManualBoost={20}
-        onQueueManualBoostChange={() => {}}
-        queueFreshBoost={5}
-        onQueueFreshBoostChange={() => {}}
-        queueBulkBoost={0}
-        onQueueBulkBoostChange={() => {}}
-        lowAdYieldAction="nothing"
-        onLowAdYieldActionChange={() => {}}
-        episodeLogRetentionDays={30}
-        onEpisodeLogRetentionDaysChange={() => {}}
-        episodeLogLevel="debug"
-        onEpisodeLogLevelChange={() => {}}
-        textRecurrenceHints={false}
-        onTextRecurrenceHintsChange={() => {}}
-      />
-      <button onClick={() => onCommit({ processNewEpisodesFirst })}>Commit</button>
-    </>
-  );
-}
-
-describe('GlobalDefaultsSection: feed refresh interval', () => {
-  it('shows the default value of 15', () => {
-    render(<Harness onCommit={() => {}} />);
-    expect((screen.getByLabelText('Feed refresh interval') as HTMLInputElement).value).toBe('15');
-  });
-
-  it('has min/max attributes of 5 and 1440', () => {
-    render(<Harness onCommit={() => {}} />);
-    const input = screen.getByLabelText('Feed refresh interval') as HTMLInputElement;
-    expect(input.min).toBe('5');
-    expect(input.max).toBe('1440');
-  });
-
-  it('commits rssRefreshIntervalMinutes after editing the field', async () => {
-    let committed: number | null = null;
-    render(<Harness onCommit={(minutes) => { committed = minutes; }} />);
-    const user = userEvent.setup();
-
-    const input = screen.getByLabelText('Feed refresh interval');
-    await user.clear(input);
-    await user.type(input, '30');
-    input.blur();
-
-    await user.click(screen.getByRole('button', { name: 'Commit' }));
-    expect(committed).toBe(30);
-  });
-
-  it('clamps a value above the max to 1440', async () => {
-    let committed: number | null = null;
-    render(<Harness onCommit={(minutes) => { committed = minutes; }} />);
-    const user = userEvent.setup();
-
-    const input = screen.getByLabelText('Feed refresh interval');
-    await user.clear(input);
-    await user.type(input, '5000');
-    input.blur();
-
-    await user.click(screen.getByRole('button', { name: 'Commit' }));
-    expect(committed).toBe(1440);
-  });
-
-  it('clamps a value below the min to 5', async () => {
-    let committed: number | null = null;
-    render(<Harness onCommit={(minutes) => { committed = minutes; }} />);
-    const user = userEvent.setup();
-
-    const input = screen.getByLabelText('Feed refresh interval');
-    await user.clear(input);
-    await user.type(input, '1');
-    input.blur();
-
-    await user.click(screen.getByRole('button', { name: 'Commit' }));
-    expect(committed).toBe(5);
-  });
-});
 
 describe('GlobalDefaultsSection: Podping notifications toggle', () => {
   it('renders off by default', () => {
@@ -220,25 +107,6 @@ describe('GlobalDefaultsSection: Podping notifications toggle', () => {
   });
 });
 
-describe('GlobalDefaultsSection: Process new episodes first toggle', () => {
-  it('renders on by default', () => {
-    render(<ProcessNewFirstHarness onCommit={() => {}} />);
-    const toggle = screen.getByRole('switch', { name: 'Process new episodes first' });
-    expect(toggle.getAttribute('aria-checked')).toBe('true');
-  });
-
-  it('commits { processNewEpisodesFirst: false } after switching off', async () => {
-    let committed: ProcessNewFirstState | null = null;
-    render(<ProcessNewFirstHarness onCommit={(payload) => { committed = payload; }} />);
-    const user = userEvent.setup();
-
-    await user.click(screen.getByRole('switch', { name: 'Process new episodes first' }));
-    await user.click(screen.getByRole('button', { name: 'Commit' }));
-
-    expect(committed).toEqual({ processNewEpisodesFirst: false });
-  });
-});
-
 describe('GlobalDefaultsSection: no Segment actions details block', () => {
   it('does not render a Segment actions details element (moved to SegmentActionsSection)', () => {
     const { container } = render(<Harness onCommit={() => {}} />);
@@ -266,14 +134,6 @@ function LowAdYieldHarness({ onCommit }: { onCommit: (payload: LowAdYieldState) 
         onMaxFeedEpisodesChange={() => {}}
         onlyExposeProcessedDefault={false}
         onOnlyExposeProcessedDefaultChange={() => {}}
-        processNewEpisodesFirst
-        onProcessNewEpisodesFirstChange={() => {}}
-        queueManualBoost={20}
-        onQueueManualBoostChange={() => {}}
-        queueFreshBoost={5}
-        onQueueFreshBoostChange={() => {}}
-        queueBulkBoost={0}
-        onQueueBulkBoostChange={() => {}}
         lowAdYieldAction={lowAdYieldAction}
         onLowAdYieldActionChange={setLowAdYieldAction}
         episodeLogRetentionDays={30}
@@ -335,14 +195,6 @@ function EpisodeLogHarness({ onCommit }: { onCommit: (payload: EpisodeLogState) 
         onMaxFeedEpisodesChange={() => {}}
         onlyExposeProcessedDefault={false}
         onOnlyExposeProcessedDefaultChange={() => {}}
-        processNewEpisodesFirst
-        onProcessNewEpisodesFirstChange={() => {}}
-        queueManualBoost={20}
-        onQueueManualBoostChange={() => {}}
-        queueFreshBoost={5}
-        onQueueFreshBoostChange={() => {}}
-        queueBulkBoost={0}
-        onQueueBulkBoostChange={() => {}}
         lowAdYieldAction="nothing"
         onLowAdYieldActionChange={() => {}}
         episodeLogRetentionDays={retentionDays}
@@ -407,14 +259,6 @@ function TextRecurrenceHintsHarness({ onCommit }: { onCommit: (payload: TextRecu
         onMaxFeedEpisodesChange={() => {}}
         onlyExposeProcessedDefault={false}
         onOnlyExposeProcessedDefaultChange={() => {}}
-        processNewEpisodesFirst
-        onProcessNewEpisodesFirstChange={() => {}}
-        queueManualBoost={20}
-        onQueueManualBoostChange={() => {}}
-        queueFreshBoost={5}
-        onQueueFreshBoostChange={() => {}}
-        queueBulkBoost={0}
-        onQueueBulkBoostChange={() => {}}
         lowAdYieldAction="nothing"
         onLowAdYieldActionChange={() => {}}
         episodeLogRetentionDays={30}
@@ -445,55 +289,5 @@ describe('GlobalDefaultsSection: Text recurrence hints toggle', () => {
     await user.click(screen.getByRole('button', { name: 'Commit' }));
 
     expect(committed).toEqual({ textRecurrenceHints: true });
-  });
-});
-
-describe('GlobalDefaultsSection: queue boosts', () => {
-  function renderBoosts(onManual = () => {}) {
-    return render(
-      <GlobalDefaultsSection
-        autoProcessEnabled={false}
-        onAutoProcessEnabledChange={() => {}}
-        rssRefreshIntervalMinutes={15}
-        onRssRefreshIntervalMinutesChange={() => {}}
-        podpingEnabled={false}
-        onPodpingEnabledChange={() => {}}
-        maxFeedEpisodes={10}
-        onMaxFeedEpisodesChange={() => {}}
-        onlyExposeProcessedDefault={false}
-        onOnlyExposeProcessedDefaultChange={() => {}}
-        processNewEpisodesFirst={true}
-        onProcessNewEpisodesFirstChange={() => {}}
-        queueManualBoost={20}
-        onQueueManualBoostChange={onManual}
-        queueFreshBoost={5}
-        onQueueFreshBoostChange={() => {}}
-        queueBulkBoost={0}
-        onQueueBulkBoostChange={() => {}}
-        lowAdYieldAction="nothing"
-        onLowAdYieldActionChange={() => {}}
-        episodeLogRetentionDays={30}
-        onEpisodeLogRetentionDaysChange={() => {}}
-        episodeLogLevel="info"
-        onEpisodeLogLevelChange={() => {}}
-        textRecurrenceHints={false}
-        onTextRecurrenceHintsChange={() => {}}
-      />
-    );
-  }
-
-  it('renders the three boost fields with their current values', () => {
-    renderBoosts();
-    expect((screen.getByLabelText('Play / Reprocess') as HTMLInputElement).value).toBe('20');
-    expect((screen.getByLabelText('New episode') as HTMLInputElement).value).toBe('5');
-    expect((screen.getByLabelText('Reprocess All') as HTMLInputElement).value).toBe('0');
-  });
-
-  it('reports an edited manual boost', () => {
-    const onManual = vi.fn();
-    renderBoosts(onManual);
-    const input = screen.getByLabelText('Play / Reprocess');
-    fireEvent.change(input, { target: { value: '30' } });
-    expect(onManual).toHaveBeenCalledWith(30);
   });
 });
