+++ frontend/src/pages/settings/GlobalDefaultsSection.tsx
@@ -16,14 +16,6 @@ interface GlobalDefaultsSectionProps {
   onMaxFeedEpisodesChange: (n: number) => void;
   onlyExposeProcessedDefault: boolean;
   onOnlyExposeProcessedDefaultChange: (enabled: boolean) => void;
-  processNewEpisodesFirst: boolean;
-  onProcessNewEpisodesFirstChange: (enabled: boolean) => void;
-  queueManualBoost: number;
-  onQueueManualBoostChange: (value: number) => void;
-  queueFreshBoost: number;
-  onQueueFreshBoostChange: (value: number) => void;
-  queueBulkBoost: number;
-  onQueueBulkBoostChange: (value: number) => void;
   lowAdYieldAction: LowAdYieldAction;
   onLowAdYieldActionChange: (action: LowAdYieldAction) => void;
   episodeLogRetentionDays: number;
@@ -45,14 +37,6 @@ function GlobalDefaultsSection({
   onMaxFeedEpisodesChange,
   onlyExposeProcessedDefault,
   onOnlyExposeProcessedDefaultChange,
-  processNewEpisodesFirst,
-  onProcessNewEpisodesFirstChange,
-  queueManualBoost,
-  onQueueManualBoostChange,
-  queueFreshBoost,
-  onQueueFreshBoostChange,
-  queueBulkBoost,
-  onQueueBulkBoostChange,
   lowAdYieldAction,
   onLowAdYieldActionChange,
   episodeLogRetentionDays,
@@ -83,85 +67,6 @@ function GlobalDefaultsSection({
           <p className="mt-2 text-sm text-muted-foreground">
             When a feed refresh discovers a new episode, queue it for processing automatically. Per-feed Auto-Process can override this.
           </p>
-        </div>
-
-        {/* Process new episodes first: fresh-episode queue boost, saves immediately */}
-        <div className="pt-4 border-t border-border">
-          <label className="flex items-center gap-3 cursor-pointer">
-            <ToggleSwitch
-              checked={processNewEpisodesFirst}
-              onChange={onProcessNewEpisodesFirstChange}
-              ariaLabel="Process new episodes first"
-            />
-            <span className="text-sm font-medium text-foreground">
-              Process new episodes first
-            </span>
-          </label>
-          <p className="mt-2 text-sm text-muted-foreground">
-            Episodes published in the last 48 hours jump ahead of queued backlog.
-          </p>
-        </div>
-
-        {/* Queue priority: how far each kind of request jumps the queue.
-            Higher number processes sooner; ties process oldest first. */}
-        <div className="pt-4 border-t border-border">
-          <p className="text-sm font-medium text-foreground mb-1">Queue priority</p>
-          <p className="text-sm text-muted-foreground mb-3">
-            How far each request type jumps the queue. Ties process oldest first.
-          </p>
-          <div className="grid grid-cols-1 sm:grid-cols-3 gap-4">
-            <div>
-              <label htmlFor="queueManualBoost" className="block text-sm text-foreground mb-2">
-                Play / Reprocess
-              </label>
-              <NumberInput
-                id="queueManualBoost"
-                value={queueManualBoost}
-                min={0}
-                max={100}
-                step={1}
-                fallback={20}
-                onCommit={onQueueManualBoostChange}
-              />
-              <p className="mt-2 text-sm text-muted-foreground">
-                Playing an unprocessed episode or reprocessing one. Keep highest so your requests skip the backlog. Default 20.
-              </p>
-            </div>
-            <div>
-              <label htmlFor="queueFreshBoost" className="block text-sm text-foreground mb-2">
-                New episode
-              </label>
-              <NumberInput
-                id="queueFreshBoost"
-                value={queueFreshBoost}
-                min={0}
-                max={100}
-                step={1}
-                fallback={5}
-                onCommit={onQueueFreshBoostChange}
-              />
-              <p className="mt-2 text-sm text-muted-foreground">
-                Published in the last 48 hours. Needs the toggle above. Default 5.
-              </p>
-            </div>
-            <div>
-              <label htmlFor="queueBulkBoost" className="block text-sm text-foreground mb-2">
-                Reprocess All
-              </label>
-              <NumberInput
-                id="queueBulkBoost"
-                value={queueBulkBoost}
-                min={0}
-                max={100}
-                step={1}
-                fallback={0}
-                onCommit={onQueueBulkBoostChange}
-              />
-              <p className="mt-2 text-sm text-muted-foreground">
-                Reprocess All and segment re-renders. Leave at 0 to keep backfills last. Default 0.
-              </p>
-            </div>
-          </div>
         </div>
 
         {/* Feed refresh interval */}
